Remove accession column in CustomiseTableContainer.
